Added soundTest app
This is the very basic app which I used to figure out how to get MATLAB to play MP3 files. This will not be used in the final app, but I am putting it here for safe keeping and just in case anyone wants to take a look at it.